expect as dispatch
