+++ netket/vqs/__init__.py
@@ -17,6 +17,7 @@ from .mc_state import MCState
 from .mc_mixed_state import MCMixedState
 
 from . import mc_expect
+from . import mc_expect_grad
 
 from . import experimental
